requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
pyarrow
boto3
pyspark==3.5.1
dbt-core==1.7.9
dbt-postgres==1.7.9
